fix_1 PR-21
