+++ run_bot.py
@@ -1,7 +1,6 @@
 import logging
 
 import dotenv
-from telegram.ext import Updater
 
 import bot
 
@@ -9,6 +8,4 @@ dotenv.load_dotenv()
 
 if __name__ == "__main__":
     logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
-    u = bot.init()
-    if isinstance(u, Updater):
-        u.idle()
+    bot.init()
